createtablewithlongpacketsent&received

+++ mss-dashboard-hal/src/main/java/pfe/cheima/connect_to_mss/Trafficforsigu.java
@@ -17,8 +17,8 @@ public class Trafficforsigu implements Serializable {
     private Integer id;
     private String siguName;
     private Date dateExec;
-    private Integer packetreceived;
-    private Integer packetsent;
+    private Long packetreceived;
+    private Long packetsent;
     private Integer SiguId;
 
     public Integer getSiguId() {
@@ -54,19 +54,19 @@ public class Trafficforsigu implements Serializable {
         this.dateExec = dateExec;
     }
 
-    public Integer getPacketreceived() {
+    public Long getPacketreceived() {
         return packetreceived;
     }
 
-    public void setPacketreceived(Integer packetreceived) {
+    public void setPacketreceived(Long packetreceived) {
         this.packetreceived = packetreceived;
     }
 
-    public Integer getPacketsent() {
+    public Long getPacketsent() {
         return packetsent;
     }
 
-    public void setPacketsent(Integer packetsent) {
+    public void setPacketsent(Long packetsent) {
         this.packetsent = packetsent;
     }
